+++ src/components/AppLocaleSelector.vue
@@ -1,7 +1,23 @@
 <script setup lang="ts">
 import { useI18n } from 'vue-i18n';
+import { watch } from 'vue';
+import { Quasar } from 'quasar';
+import { quasarLangList } from 'src/utils/quasar-lang';
 
 const { locale, availableLocales } = useI18n({ useScope: 'global' });
+
+watch(locale, () => {
+  // Load Quasar Language Pack
+  if (locale.value === 'en') {
+    quasarLangList['../../node_modules/quasar/lang/en-US.mjs']().then((lang) => {
+      Quasar.lang.set(lang.default);
+    });
+  } else {
+    quasarLangList[`../../node_modules/quasar/lang/${locale.value}.mjs`]().then((lang) => {
+      Quasar.lang.set(lang.default);
+    });
+  }
+}, { immediate: true });
 </script>
 
 <template>